# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LarisSama Backend API — Draft
  version: 0.1.6-draft
  description: >-
    Status handoff 2026-10-08: 37 operationId tersedia untuk integrasi alur utama; GET data tenant oleh superadmin wajib menyertakan `warung_id` yang dipilih.
    Seluruh aksi transaksi juga tersedia bagi superadmin dengan `warung_id` wajib pada body. Request tenant menggunakan warung dari identitas dan menolak field selector.
    payPenjualan sudah melalui focused conformance
    untuk integrasi alur utama, termasuk pendaftaran owner, persetujuan admin,
    langganan 30 hari, serta operasi koreksi/pembatalan/retur.
    Pengujian edge-case lanjutan dijadwalkan setelah integrasi frontend dimulai;
    x-deferred-verification per operationId menjelaskan batasnya. x-contract-status
    adalah status handoff terkini. User menyetujui baseline wire D13: /api/v1,
    ID/decimal string, envelope data/meta, page integer >= 1 tanpa batas maksimum dan per_page 1–100, sort
    allowlist, dan error code/message/errors/request_id. Slice transaksi dan auth
    memiliki feature test HTTP pada MySQL 8.0.40 untuk sebagian create/read/report,
    tenant scope, snapshot, retry, atomicity, batas DST, periode kosong/invalid,
    login/me, role context, status akun, expiry/logout, rate limit, provisioning
    admin warung/user tenant, katalog, dan transaksi/laporan memiliki acceptance
    test parsial. Response runtime auth, admin, katalog, dan transaksi/laporan dicocokkan
...
    kontrak live.
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /penjualans:
    get:
      operationId: listPenjualans
      tags:
      - Penjualan
      summary: Riwayat Penjualan
      description: >-
        READY_FOR_FRONTEND untuk alur inti; extended edge-case tercatat untuk susulan. Setiap item berasal
        dari menu terdaftar; owner, manager, dan kasir membaca seluruh penjualan pada warung sendiri.
        TRANSACTION-CONFORMANCE-001 mencocokkan response 200 dengan schema. TRANSACTION-QUERY-CONFORMANCE-001 mencocokkan page/per_page/sort/status pada query sukses terpilih.
        ACCESS-PAGINATION-MAX-CONFORMANCE-001 menguji per_page=100 lewat HTTP dan memeriksa meta response;
        ACCESS-PAGINATION-OVERFLOW-CONFORMANCE-001 membuktikan per_page=101 ditolak dengan HTTP 422, schema response 422,
        dan error field per_page; nilai/filter lain, status lain, dan gate penuh masih terbuka.
        ACCESS-PAGINATION-LOWER-BOUND-CONFORMANCE-001 membuktikan page=0 dan per_page=0 ditolak dengan HTTP 422,
        schema response 422, dan error field terkait.
        ACCESS-PAGINATION-TYPE-CONFORMANCE-001 membuktikan page=abc dan per_page=abc ditolak dengan HTTP 422,
        schema response 422, dan error field terkait.
        TRANSACTION-LIST-QUERY-REJECTION-001 membuktikan status sale di luar enum dan tanggal malformed/range terbalik
        pada kedua list ditolak dengan HTTP 422 schema-conformant dan error field yang tepat.
        TRANSACTION-LIST-QUERY-PAIR-REQUIRED-001 membuktikan filter date_from/date_to parsial ditolak 422; keduanya wajib dipasangkan.
        TRANSACTION-LIST-LOCAL-PERIOD-CONFORMANCE-001 membuktikan batas hari Asia/Jakarta, dan TRANSACTION-LIST-DST-CONFORMANCE-001 membuktikan batas hari DST 23 jam di America/New_York pada kedua list.
        SUPERADMIN-TRANSACTION-LIST-RBAC-CONFORMANCE-001 membuktikan hasil 403 historis untuk superadmin kini digantikan akses baca dengan warung_id wajib; `OWNER-TENANT-ACCESS-CONFORMANCE-001` juga menguji owner membaca transaksi di tenant sendiri dan menerima 404 untuk transaksi tenant lain.
        API-END-TO-END-READBACK-001 menguji manager membaca list sale sesudah kasir membuat sale lewat API.
        D18 menggantikan batas lama riwayat kasir sendiri: semua role operasional membaca transaksi satu warung; filter
        status_pembayaran memisahkan antrean pending dan lunas. Alur lintas pencatat diuji pada SalesOrderPaymentApiTest.
      x-contract-status: READY_FOR_FRONTEND
      x-deferred-verification: "Extended error/status and edge-case conformance will be completed after frontend integration begins; see docs/api/README.md."
      x-implementation-status: DONE
      x-candidate-roles:
      - owner
      - manager
      - kasir
      - superadmin
      x-blocked-by:
      - D02
      - D13
      responses:
        '200':
          description: Kandidat hasil sukses; contoh sintetis, bukan data server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanSummaryListResponse'
              example:
                data:
                - id: '501'
                  warung_id: '1'
                  user_id: '11'
                  no_transaksi: PJ-CONTOH-001
                  tanggal: '2026-10-04T03:00:00Z'
                  subtotal: '35000.00'
                  diskon: '2000.00'
                  total: '33000.00'
                  bayar: '50000.00'
                  kembalian: '17000.00'
                  metode_pembayaran: cash
                  status: selesai
                  catatan: null
                  created_at: '2026-10-04T03:00:00Z'
                  updated_at: '2026-10-04T03:00:00Z'
                meta:
                  page: 1
                  per_page: 20
                  total: 1
                  last_page: 1
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/WarungReadScope'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - -tanggal
          - tanggal
          default: -tanggal
        description: Allowlist urutan; backend menambah id sebagai tie-breaker agar stabil.
      - name: date_from
        in: query
        schema:
          $ref: '#/components/schemas/Date'
        description: Awal hari lokal inklusif (D08); kirim bersama date_to. Respons 422 pada date_from bila batas awal UTC berada di luar rentang MySQL DATETIME tahun 1000–9999.
        required: false
      - name: date_to
        in: query
        schema:
          $ref: '#/components/schemas/Date'
        description: Tanggal akhir lokal inklusif (D08); date_from <= date_to. Respons 422 pada date_to bila batas akhir-eksklusif UTC berada di luar rentang MySQL DATETIME tahun 1000–9999.
        required: false
      - name: status
        in: query
        schema:
          type: string
          enum:
          - menunggu_pembayaran
          - selesai
          - batal
          - diretur_sebagian
          - diretur_penuh
        description: Filter status; tanpa filter seluruh status penjualan ditampilkan.
      - name: status_pembayaran
        in: query
        schema:
          type: string
          enum: [belum_lunas, lunas]
        description: Filter status pembayaran. `belum_lunas` hanya menampilkan pesanan yang masih dapat dibayar; pesanan batal tidak ikut antrean pembayaran.
    post:
      operationId: createPenjualan
      tags:
      - Penjualan
      summary: Catat Penjualan
      description: >-
        Menyimpan pesanan dan rincian atomik; warung/user/nomor/total dari backend. Jika `bayar` dan
        `metode_pembayaran` tidak dikirim, pesanan dibuat dengan status menunggu_pembayaran dan dapat diedit
        sampai lunas. Untuk penjualan langsung di kasir, pasangan pembayaran boleh dikirim dan transaksi langsung lunas.
        Nama pelanggan bersifat teks bebas dan opsional. `no_transaksi` dari backend menjadi nomor pada nota cetak.
        Kirim Idempotency-Key untuk retry: payload sama me-replay hasil awal dengan HTTP 201, payload berbeda
        memakai key sama menghasilkan 409. D05 diputuskan: harga menu > 0, qty > 0 hingga dua desimal, diskon
        rincian/header tidak lebih besar dari subtotal, cash boleh lebih dari total, QRIS/transfer harus tepat, dan
        pembulatan half-up per rincian. READY_FOR_FRONTEND untuk alur create inti; variasi edge-case tercatat untuk susulan.
        Setiap item wajib berasal dari menu aktif dalam kategori aktif. `TRANSACTION-INACTIVE-CATALOG-SALE-CONFORMANCE-001`
        membuktikan menu nonaktif atau kategori nonaktif ditolak 422 pada rincian tanpa write, sedangkan katalog aktif menerima sale.
        TRANSACTION-CONFORMANCE-001 mencocokkan response 201/403/409/422;
        TRANSACTION-REQUEST-CONFORMANCE-001 mencocokkan body/header pada request create/replay/conflict terpilih dan
        tenant-crossing 422 schema-valid (menu_id berupa string). SUPERADMIN-TRANSACTION-RBAC-CONFORMANCE-001 menguji
        Superadmin wajib mengirim warung_id pada body; tanpa selector request mendapat 422 dan scope silang mendapat 404.
        ROLE-TRANSACTION-CREATE-RBAC-CONFORMANCE-001 mencatat aturan lama; D18 sekarang memberi hak create kepada manager.
        Superadmin mengirim warung_id target pada body; field ini dilarang untuk user tenant. SALE-DECIMAL-EXACT-CONFORMANCE-001 membuktikan harga 17.25 × qty 3.00
        menghasilkan subtotal/total 51.75 dan kembalian 8.25 sebagai decimal eksak di response dan MySQL.
        Query, body/status lain, serta gate penuh masih terbuka.
      x-contract-status: READY_FOR_FRONTEND
      x-deferred-verification: "Extended error/status and edge-case conformance will be completed after frontend integration begins; see docs/api/README.md."
      x-implementation-status: DONE
      x-candidate-roles:
      - owner
      - manager
      - kasir
      - superadmin
      x-blocked-by:
      - D02
      - D13
      - D09
      responses:
        '201':
          description: Kandidat hasil sukses; contoh sintetis, bukan data server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanResponse'
              examples:
                menu:
                  value:
                    data:
                      id: '501'
                      warung_id: '1'
                      user_id: '11'
                      no_transaksi: PJ-CONTOH-001
                      tanggal: '2026-10-04T03:00:00Z'
                      subtotal: '35000.00'
                      diskon: '2000.00'
                      total: '33000.00'
                      bayar: '50000.00'
                      kembalian: '17000.00'
                      metode_pembayaran: cash
                      status: selesai
                      catatan: null
                      created_at: '2026-10-04T03:00:00Z'
                      updated_at: '2026-10-04T03:00:00Z'
                      rincian:
                      - id: '5001'
                        penjualan_id: '501'
                        menu_id: '1001'
                        nama_menu: Nasi
                        harga: '15000.00'
                        qty: '2.00'
                        diskon: '0.00'
                        subtotal: '30000.00'
                        catatan: null
                        created_at: '2026-10-04T03:00:00Z'
                        updated_at: '2026-10-04T03:00:00Z'
                      - id: '5002'
                        penjualan_id: '501'
                        menu_id: '1002'
                        nama_menu: Teh
                        harga: '5000.00'
                        qty: '1.00'
                        diskon: '0.00'
                        subtotal: '5000.00'
                        catatan: null
                        created_at: '2026-10-04T03:00:00Z'
                        updated_at: '2026-10-04T03:00:00Z'
                      riwayat_koreksi: []
                      riwayat_retur: []
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/IdempotencyConflict409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^\S(?:.*\S)?$'
        description: Key unik per warung, user, dan endpoint selama 7 hari sejak request pertama; 1–255 karakter tanpa whitespace di awal/akhir. Setelah expiry, key lama tidak me-replay respons dan pemakaian ulang diproses sebagai request baru. Metadata lama dibersihkan saat transaksi request baru commit; jika request ditolak aturan bisnis dan rollback, metadata expired dapat tetap tersimpan tetapi tidak berlaku untuk replay.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenjualanCreate'
            examples:
              penjualan_langsung_lunas:
                value:
                  tanggal: '2026-10-04T03:00:00Z'
                  diskon: '2000.00'
                  bayar: '50000.00'
                  metode_pembayaran: cash
                  rincian:
                  - menu_id: '1001'
                    qty: '2.00'
                  - menu_id: '1002'
                    qty: '1.00'
              pesanan_belum_lunas:
                value:
                  tanggal: '2026-10-07T10:00:00+07:00'
                  nama_pelanggan: 'Andi'
                  rincian:
                  - menu_id: '1001'
                    qty: '2.00'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Mekanisme terpilih: Sanctum bearer token; `laravel/sanctum` v4.3.3 dan migration token sudah terpasang.
        Token kedaluwarsa setelah 30 hari; logout mencabut token aktif. Login dibatasi 5 percobaan per menit per kombinasi username+IP.
        Route auth sudah dibuat tetapi belum lulus verifikasi aplikasi/kontrak. Personal access token opaque; CORS, HTTPS, dan detail deployment masih D02.'
  parameters:
    WarungReadScope:
      name: warung_id
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Id'
      description: Wajib untuk superadmin pada seluruh GET data tenant dan harus menunjuk ID warung yang ada (ID tidak ada menghasilkan 422). Untuk detail, resource di luar warung pilihan menghasilkan 404. Pengguna tenant dilarang mengirim parameter ini; scope mereka selalu berasal dari bearer token. Selector baca tidak memberi wewenang mutasi katalog atau user tenant. Aksi tulis transaksi memakai warung_id di body.
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Halaman bilangan bulat positif tanpa batas maksimum. Halaman setelah last_page mengembalikan data kosong dan mempertahankan total hasil filter. Nilai di atas rentang integer PHP tetap diterima dan dikirim sebagai token integer JSON eksak; bukti keenam GET list ada pada UNBOUNDED-PAGINATION-PAGE-CONFORMANCE-001.
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Error400:
      description: JSON request tidak dapat dibaca.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: BAD_REQUEST
            message: JSON request tidak dapat dibaca.
            errors: {}
            request_id: req-contoh-001
    Error401:
      description: Login diperlukan atau kredensial/token tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: UNAUTHENTICATED
            message: Login diperlukan atau kredensial/token tidak valid.
            errors: {}
            request_id: req-contoh-001
    Error403:
      description: Role tidak diizinkan; user/warung nonaktif atau masa aktif tidak memenuhi D03.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: FORBIDDEN
            message: Role tidak diizinkan; user/warung nonaktif atau masa aktif tidak memenuhi D03.
            errors: {}
            request_id: req-contoh-001
    Error404:
      description: Resource tidak ada dalam scope yang boleh diakses.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: NOT_FOUND
            message: Resource tidak ada dalam scope yang boleh diakses.
            errors: {}
            request_id: req-contoh-001
    Error409:
      description: Konflik keunikan atau state yang sudah disepakati.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: CONFLICT
            message: Konflik keunikan atau state yang sudah disepakati.
            errors: {}
            request_id: req-contoh-001
    IdempotencyConflict409:
      description: Idempotency-Key yang sama sudah tercatat untuk payload berbeda.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: IDEMPOTENCY_KEY_REUSED
            message: Idempotency-Key sudah digunakan untuk payload berbeda.
            errors: {}
            request_id: req-contoh-001
    Error422:
      description: Field tidak valid, referensi tenant tidak sah, nominal/rentang tanggal salah.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: VALIDATION_ERROR
            message: Field tidak valid, referensi tenant tidak sah, nominal/rentang tanggal salah.
            errors:
              rincian.0.nama_item:
              - Wajib diisi.
            request_id: req-contoh-001
    Error429:
      description: Terlalu banyak request menurut batas yang ditetapkan D02.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: RATE_LIMITED
            message: Terlalu banyak request menurut batas yang ditetapkan D02.
            errors: {}
            request_id: req-contoh-001
    Error500:
      description: Kesalahan server; response tidak mengungkap detail SQL/rahasia.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INTERNAL_ERROR
            message: Kesalahan server; response tidak mengungkap detail SQL/rahasia.
            errors: {}
            request_id: req-contoh-001
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      description: ID BIGINT dikirim sebagai string agar tidak kehilangan presisi di JavaScript.
    Money:
      type: string
      pattern: ^(0|[1-9][0-9]{0,12})\.[0-9]{2}$
      description: Decimal string eksak nonnegatif dengan dua angka pecahan sesuai baseline D05; jangan memakai float. Batas saat ini mengikuti DECIMAL(15,2).
...
    Timestamp:
      type: string
      format: date-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|[+-](?:[01]\d|2[0-3]):[0-5]\d)$'
      description: 'Profil RFC3339: tahun 1000–9999, jam 00–23, menit/detik 00–59, zona Z atau offset ±HH:MM. Pada create transaksi, offset diubah ke UTC dan instant harus berada dalam rentang MySQL DATETIME (1000-01-01T00:00:00Z sampai 9999-12-31T23:59:59Z); instant di luar rentang ditolak. Future transaction instant ditolak; backdate diterima. Pecahan detik pada request create transaksi diterima tetapi dipotong ke presisi detik DATETIME, tanpa pembulatan. Leap second tidak didukung. Timestamp disimpan/dikirim UTC; tampilan/periode memakai zona IANA warung. Zona kosong/invalid menolak akses.'
...
    Date:
      type: string
      format: date
    Note:
      type:
      - string
      - 'null'
      description: Catatan bebas, null jika tidak diisi.
...
    PenjualanCreate:
      type: object
      additionalProperties: false
      properties:
        warung_id:
          $ref: '#/components/schemas/Id'
        tanggal:
          $ref: '#/components/schemas/Timestamp'
        diskon:
          $ref: '#/components/schemas/Money'
        nama_pelanggan:
          type: [string, 'null']
          maxLength: 150
        bayar:
          $ref: '#/components/schemas/Money'
        metode_pembayaran:
          type: string
          enum:
          - cash
          - qris
          - transfer
          description: 'D05 diputuskan: cash boleh melebihi total; QRIS dan transfer harus sama persis dengan total.'
        catatan:
          $ref: '#/components/schemas/Note'
        rincian:
          type: array
          items:
            $ref: '#/components/schemas/SaleMenuLineInput'
          minItems: 1
      required:
      - tanggal
      - rincian
      description: >-
        Tanpa bayar dan metode_pembayaran, buat pesanan belum lunas. Sertakan keduanya hanya untuk penjualan
        langsung lunas. Nama pelanggan adalah teks bebas opsional. Diskon yang tidak dikirim dianggap 0.00;
        Untuk superadmin, warung_id wajib menunjuk warung target; user tenant dilarang mengirimnya. Subtotal/total/kembalian/identitas pelaku/no_transaksi/status ditetapkan backend.
...
    PenjualanSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        warung_id:
          $ref: '#/components/schemas/Id'
        user_id:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        created_by_superadmin_id:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        no_transaksi:
          type: string
          minLength: 1
          maxLength: 50
        nama_pelanggan:
          type: [string, 'null']
          maxLength: 150
        tanggal:
          $ref: '#/components/schemas/Timestamp'
        subtotal:
          $ref: '#/components/schemas/Money'
        diskon:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        bayar:
          $ref: '#/components/schemas/Money'
        kembalian:
          $ref: '#/components/schemas/Money'
        metode_pembayaran:
          type: [string, 'null']
          enum:
          - cash
          - qris
          - transfer
          description: 'D05 diputuskan: cash boleh melebihi total; QRIS dan transfer harus sama persis dengan total.'
        status:
          type: string
          enum:
          - menunggu_pembayaran
          - selesai
          - batal
          - diretur_sebagian
          - diretur_penuh
        status_pembayaran:
          type: string
          enum: [belum_lunas, lunas]
        dibayar_pada:
          anyOf:
          - $ref: '#/components/schemas/Timestamp'
          - type: 'null'
        pembayaran_user_id:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        pembayaran_superadmin_id:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        catatan:
          $ref: '#/components/schemas/Note'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required:
      - id
      - warung_id
      - user_id
      - no_transaksi
      - tanggal
      - subtotal
      - diskon
      - total
      - bayar
      - kembalian
      - metode_pembayaran
      - status
      - catatan
      - created_at
      - updated_at
...
    Pagination:
      type: object
      additionalProperties: false
      properties:
        page:
          type: integer
          minimum: 1
          description: Nomor halaman positif tanpa batas maksimum; dikirim sebagai integer JSON eksak.
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        last_page:
          type: integer
          minimum: 1
      required:
      - page
      - per_page
      - total
      - last_page
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        request_id:
          type: string
      required:
      - code
      - message
      - errors
      - request_id
...
    PenjualanSummaryListResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PenjualanSummary'
        meta:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      - meta
    PenjualanResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/Penjualan'
      required:
      - data
...